backend/requirements.txt
fastapi
uvicorn
torch
sentencepiece
python-dotenv
numpy
